# Improve chatlist scrolling performance