--- a/scss/chat/_chat-list-item.scss
+++ b/scss/chat/_chat-list-item.scss
@@ -17,6 +17,13 @@ $color-chat-list-item-avatar-text: var(--avatarLabelColor);
   &:hover {
     background-color:#bbb;
   }
+  &.skeleton {
+    background-repeat: no-repeat;
+    background-image: radial-gradient(circle 24px, grey 99%, transparent 0),linear-gradient(grey 40px, transparent 0), linear-gradient(grey 40px, transparent 0), linear-gradient(grey 40px, transparent 0);
+    background-size: 56px 56px, 200px 18px, 300px 16px, 40px 16px;
+    background-position: 6px, 67px 12px, 67px 35px, 96% 12px;
+    opacity: 0.1;
+  }
 }
 
 .chat-list-item--has-unread {
@@ -206,3 +213,4 @@ $color-chat-list-item-avatar-text: var(--avatarLabelColor);
     background-color: #bbb;
   } 
 }
+