--- a/scss/chat/_chat-list.scss
+++ b/scss/chat/_chat-list.scss
@@ -10,37 +10,5 @@
 }
 
 .chatlist-item-skeleton {
-  opacity: 0.1;
-
-  & .bone {
-    background: grey;
-    border-radius: 3pt;
-    color: grey;
-  }
-
-  .skeleton-avatar {
-    position: relative;
-    height: 48px;
-    width: 48px;
-    margin-top: 8px;
-    margin-bottom: 8px;
-    border-radius: 24px;
-    background-color: grey;
-  }
-
-  animation: breath 2s infinite;
-}
-
-@keyframes breath {
-  0% {
-    opacity: 0.1;
-  }
-
-  50% {
-    opacity: 0.12;
-  }
-
-  100% {
-    opacity: 0.1;
-  }
+  border-bottom: 1px solid #ddd;
 }