--- a/src/renderer/components/chat/ChatListItem.js
+++ b/src/renderer/components/chat/ChatListItem.js
@@ -74,7 +74,7 @@ export const PlaceholderChatListItem = React.memo((props) => {
     className={classNames(
       'chat-list-item',
       props.className,
-      'chatlist-item-skeleton'
+      'skeleton'
     )}
   />
 })