--- a/test/integration/index.js
+++ b/test/integration/index.js
@@ -22,7 +22,7 @@ test('Bad mail address results in error message', async (t) => {
       .setValue('#mail_pw', 'bar')
       .click('button[type=\'submit\']')
 
-    await setup.wait(1500)
+    await setup.wait(1000)
     const text = await app.client.getText('.user-feedback')
     t.equal(text, 'Bad email address.', 'Mail validation error is shown')
     setup.endTest(app, t)