This improves the chatlist scrolling performance by splitting some heavy backend methods into some async callback "chunks" which blocks the event loop only for the time each chunk takes and allows other things to happen in between instead of blocking for the whole time. This is important for any ipc events getting queued up in between and unblocking the renderer process. This is still not perfect but feels snappier and smoother as before. Besides this, this also tries to fetch the deltachat-core chatList datastructure only once.

# Benchmarks

### improved
```
[BENCH] getChatListItemByIds 26: 132.433ms
[BENCH] getChatListItemByIds 56: 253.712ms
[BENCH] getChatListItemByIds 16: 83.381ms
[BENCH] getChatListItemByIds 1: 18.506ms
[BENCH] getChatListItemByIds 5: 36.413ms
[BENCH] getChatListItemByIds 8: 49.686ms
[BENCH] getChatListItemByIds 8: 49.323ms
[BENCH] getChatListItemByIds 1: 16.919ms
[BENCH] getChatListItemByIds 1: 16.367ms
[BENCH] getChatListItemByIds 34: 161.838ms
[BENCH] getChatListItemByIds 6: 39.748ms
```

Average: 858,326ms/162=5,26ms

### old
```
[BENCH] getChatListItemByIds 26: 395.527ms
[BENCH] getChatListItemByIds 34: 492.897ms
[BENCH] getChatListItemByIds 24: 346.927ms
[BENCH] getChatListItemByIds 12: 179.940ms
[BENCH] getChatListItemByIds 33: 476.247ms
[BENCH] getChatListItemByIds 22: 325.160ms
[BENCH] getChatListItemByIds 10: 147.971ms
[BENCH] getChatListItemByIds 20: 297.857ms
[BENCH] getChatListItemByIds 10: 147.255ms
```

Average: 12809,78ms/191=14,710895288ms